+++ test/src/cereal.c
@@ -23,10 +23,10 @@ int main(int argc, const char **argv)
 
 	b = pack_sm(eu1, buf);
 	b += pack_sm_ent(eu1->update, &buf[b]);
-	printf("packed %d bytes\n", b);
+	printf("packed %ld bytes\n", (long)b);
 	b = unpack_sm(eu2, buf);
 	b += unpack_sm_ent(eu2->update, &buf[b]);
-	printf("unpacked %d bytes\n", b);
+	printf("unpacked %ld bytes\n", (long)b);
 
 	eu = eu1->update;
 	printf("> %d | %d | %d | %d\n", eu1->type, eu->id, eu->pos.x, eu->pos.y);
@@ -52,14 +52,14 @@ int main(int argc, const char **argv)
 	cm1 = cm_create(client_message_action, &move);
 	b = pack_cm(cm1, buf);
 	b += pack_cm_action(cm1->update, &buf[b]);
-	printf("packed %d bytes\n", b);
+	printf("packed %ld bytes\n", (long)b);
 	struct cm_action *cma = cm1->update;
 	printf("cm_act: type: %d, (%d, %d), %d\n", cma->type, cma->range.center.x, cma->range.center.y, cma->range.r);
 
 	cm2 = cm_create(client_message_action, NULL);
 	b = unpack_cm(cm2, buf);
 	b += unpack_cm_action(cm2->update, &buf[b]);
-	printf("unpacked %d bytes\n", b);
+	printf("unpacked %ld bytes\n", (long)b);
 	cma = cm2->update;
 	printf("cm_act: type: %d, (%d, %d), %d\n", cma->type, cma->range.center.x, cma->range.center.y, cma->range.r);
